+++ src/lib.rs
@@ -86,7 +86,6 @@ use style::{StyledTextBox, TextBoxStyle};
 pub mod prelude {
     pub use crate::{
         style::{TextBoxStyle, TextBoxStyleBuilder},
-        utils::font_ext::FontExt,
         TextBox,
     };