+++ src/style/mod.rs
@@ -5,8 +5,9 @@ pub mod builder;
 
 use crate::{
     alignment::TextAlignment,
+    parser::{Parser, Token},
     rendering::{StateFactory, StyledTextBoxIterator},
-    utils::rect_ext::RectExt,
+    utils::{font_ext::FontExt, rect_ext::RectExt},
     TextBox,
 };
 pub use builder::TextBoxStyleBuilder;
@@ -39,6 +40,63 @@ where
             text_style: TextStyle::new(font, text_color),
             alignment,
         }
+    }
+
+    /// Measures text height when rendered using a given width.
+    #[inline]
+    #[must_use]
+    pub fn measure_text(&self, text: &str, max_width: u32) -> u32 {
+        let line_count = text
+            .lines()
+            .map(|line| {
+                let mut current_rows = 1;
+                let mut total_width = 0;
+                for token in Parser::parse(line) {
+                    match token {
+                        Token::Word(w) => {
+                            let mut word_width = 0;
+                            for c in w.chars() {
+                                let width = F::total_char_width(c);
+                                if total_width + word_width + width <= max_width {
+                                    // letter fits, letter is added to word width
+                                    word_width += width;
+                                } else {
+                                    // letter (and word) doesn't fit this line, open a new one
+                                    current_rows += 1;
+                                    if total_width == 0 {
+                                        // first word gets a line break in current pos
+                                        word_width = width;
+                                        total_width = 0;
+                                    } else {
+                                        // other words get wrapped
+                                        word_width += width;
+                                        total_width = 0;
+                                    }
+                                }
+                            }
+
+                            total_width += word_width;
+                        }
+
+                        Token::Whitespace(n) => {
+                            let width = F::total_char_width(' ');
+                            for _ in 0..n {
+                                if total_width + width <= max_width {
+                                    total_width += width;
+                                } else {
+                                    current_rows += 1;
+                                    total_width = width;
+                                }
+                            }
+                        }
+
+                        Token::NewLine => {}
+                    }
+                }
+                current_rows
+            })
+            .sum::<u32>();
+        line_count * F::CHARACTER_SIZE.height
     }
 }
 
@@ -115,5 +173,36 @@ where
     #[must_use]
     fn size(&self) -> Size {
         RectExt::size(self.text_box.bounds)
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::builder::TextBoxStyleBuilder;
+    use embedded_graphics::{fonts::Font6x8, pixelcolor::BinaryColor};
+
+    #[test]
+    fn test_measure_height() {
+        let data = [
+            ("", 0, 0),
+            ("word", 4 * 6, 8), // exact fit into 1 line
+            ("word", 4 * 6 - 1, 16),
+            ("word", 2 * 6, 16), // exact fit into 2 lines
+            ("word\nnext", 50, 16),
+            ("verylongword", 50, 16),
+            ("some verylongword", 50, 24),
+            ("1 23456 12345 61234 561", 36, 40),
+        ];
+        let textbox_style = TextBoxStyleBuilder::new(Font6x8)
+            .text_color(BinaryColor::On)
+            .build();
+        for (text, width, expected_height) in data.iter() {
+            let height = textbox_style.measure_text(text, *width);
+            assert_eq!(
+                height, *expected_height,
+                "Height of \"{}\" is {} but is expected to be {}",
+                text, height, expected_height
+            );
+        }
     }
 }